% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addHeadingsFCL.R
\name{addHeadingsFCL}
\alias{addHeadingsFCL}
\title{addHeadingsFCL}
\usage{
addHeadingsFCL(measuredItemFCL)
}
\description{
Function to add headings to fcl codes in order to have 4 digits fcl codes
}

