% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddViolation.R
\name{AddViolation}
\alias{AddViolation}
\title{Add violation to faosws}
\usage{
AddViolation(changeset, violationtable)
}
\arguments{
\item{changeset}{\link{Changeset} object.}

\item{violationtable.}{data.table with the following columns:
\itemize{
  \item __id
  \item column
  \item type
  \item gravity
  \item message}}
}
\description{
To mark a row as valid, just return NA for everything except ID
}

