% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BlockMetadata.R
\docType{class}
\name{BlockMetadata-class}
\alias{BlockMetadata}
\alias{BlockMetadata-class}
\title{Block Metadata Class}
\arguments{
\item{blockId}{Numeric value that holds the unique ID of the block metadata.
The GetBlockMetadata call returns this information in order to allow update
operation on block metadata, by specifying an existing ID in a
SaveBlockMetadata call.  Omitting this information in a SaveBlockMetadata
call will create a new block metadata.}

\item{selection}{A list of instances of the Dimension class.}

\item{metadata}{An instance of the Metadata class.}
}
\description{
This S4 object represents a single block metadata.
}

