% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockMetadataSet.R
\docType{class}
\name{BlockMetadataSet-class}
\alias{BlockMetadataSet}
\alias{BlockMetadataSet-class}
\title{The Block Metadata Set Class}
\description{
The Block Metadata Set Class
}
\section{Slots}{

\describe{
\item{\code{keyDefinitions}}{List of dimensions that characterize the underlying
dataset.  Each list element has to be an instance of the S4 class
KeyDefinition.}

\item{\code{blockMetadata}}{A list of BlockMetaData objects.}
}}

