% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModifiedCells.R
\name{.buildModifiedCells}
\alias{.buildModifiedCells}
\title{Build Modified Cells}
\usage{
.buildModifiedCells(dimensions, coordinates)
}
\arguments{
\item{coordinates}{}
}
\description{
Build Modified Cells
}

