% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Changeset.R
\name{Changeset}
\alias{Changeset}
\alias{Finalise}
\alias{Finalize}
\alias{Finalize}
\alias{ExtractValidationIDs}
\title{Define changeset object}
\usage{
Changeset(table, type = "write")

Finalize(changeset)

ExtractValidationIDs(changeset)
}
\arguments{
\item{table}{character. Name of Datatable on server. Use 
\code{\link{FetchDatatableConfig}} to get a list of tables on the server.}

\item{type}{character. API to write to. At the moment:
\itemize{
\item write
\item validation}}

\item{changeset}{Changeset object}
}
\value{
\code{Changeset} returns a Changeset object that can be used with the
  \code{\link{SaveDatatable}} class of functions.
}
\description{
Changeset is called to initiate a Changeset object, an object which tracks 
the rows requests sent to it (be they \link[=SaveDatatable]{insert, modify, delete} or \link[=SaveValidation]{validate}). When it 
accumulates enough requests (by default, 5000 rows), it sends that chunk to 
the server. To finalise your script, use \code{Finalize} or \code{Finalise}
with the changeset object to indicate that there are no more rows to add and
that the process of writing to a table is finished.

Changeset objects with the validation type store the IDs that have been sent 
to them. This includes IDs. A changeset object with type validation will not 
allow any rows already sent to be revalidated. The function
\code{FetchValidationIDs} retrieves any IDs stored in this Changeset.
}
\examples{
\dontrun{
ch <- Changeset("world_bank_climate_data_test")
}
}
