% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDebug.R
\name{CheckDebug}
\alias{CheckDebug}
\title{Check if the script is being run on Rserve or being debugged}
\usage{
CheckDebug()
}
\description{
This functions is a small utility function which allows checking if the 
script is running on the server or being debugged by a developer. The server
has an environment variable called R_DEBUG_MODE which is present if the 
script is running there.
}
\details{
This function thus returns TRUE when not on the server and FALSE when on the
server.
}

