% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetData.R
\name{cleanMetadata}
\alias{cleanMetadata}
\title{Clean metadata}
\usage{
cleanMetadata(metadata)
}
\arguments{
\item{metadata}{The list created by the PostRestCall.}
}
\value{
A data.table object containing the metadata values extracted
from the list.
}
\description{
This function takes a metadata object as created by a call to PostRestCall
and restructures it into a data.table object.
}

