% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariablesAdd1.R
\docType{data}
\name{CountryGroup}
\alias{CountryGroup}
\title{Part of the FAO Loss Module}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 247 rows and 18 columns.}
\usage{
CountryGroup
}
\description{
Part of the FAO Loss Module
}
\author{
Alicia English
}
\keyword{datasets}
