% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DatasetKey.R
\docType{class}
\name{DatasetKey-class}
\alias{DatasetKey}
\alias{DatasetKey-class}
\title{Dataset Key Class}
\arguments{
\item{domain}{A character value specifying the domain of interest, such as
"agriculture".}

\item{dataset}{A character value specifying the dataset within the domain,
such as "agriculture".}

\item{dimensions}{A list of objects, each of class Dimension.}

\item{sessionId}{The ID of the session from which data should be accessed.
  If NULL, the database is used directly.}
}
\description{
This class provides a mechanism for pulling data from the SWS.  You must
pass a DatasetKey object into the GetData function in order to read data
from the SWS.
}
\note{
Many of these variables (domain, dataset, and dimensions) can be found
  by examining some of the swsContext objects which are created in a
  GetTestEnvironment call (in a debug session).  Moreover, after executing
  GetTestEnvironment, the swsContext.datasets object will contain a list of
  objects of type DatasetKey.  Additionally, information about these objects
  can be found at the following url:
  http://hqlqasws1.hq.un.fao.org:8080/dataset_configuration.html
}
\examples{
\dontrun{
dim1 = Dimension(name = "geographicAreaM49", keys = c("12", "24", "204"))
dim2 = Dimension(name = "measuredElement", keys = c("51", "5510", "5607"))
dim3 = Dimension(name = "measuredItemCPC", keys = c("0111", "0112", "0113"))
dim4 = Dimension(name = "timePointYears", keys = as.character(1960:2010))

key = DatasetKey(domain = "agriculture", dataset = "agriculture",
                 dimensions = list(dim1, dim2, dim3, dim4))

GetData(key = key)
}
}
\seealso{
\code{\link{Dimension}}, \code{\link{GetData}}
}

