% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DtDimension.R
\docType{class}
\name{DtDimension-class}
\alias{DtDimension-class}
\alias{DtDimension}
\title{DtDimension class}
\description{
This class is for specifying Datatable columns
}
\section{Slots}{

\describe{
\item{\code{id}}{character. Column name.}

\item{\code{ascending}}{logical. Should this column be ascending (TRUE) or
descending (FALSE) order?}

\item{\code{value}}{For character columns, specify a single filter value. More
complex filtering can be placed in the where clause of
\code{\link{ReadDatatable}}}
}}

\examples{
mapply(new, id=letters[1:3], value=1:3, MoreArgs = list(Class="DtDimension"))

}
