% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faosws-package.R
\docType{package}
\name{faosws-package}
\alias{faosws}
\alias{faosws-package}
\title{Tools providing access to the FAO Statistical Working System}
\description{
This package allows access to the FAO SWS via its API when the appropriate 
client certificates are present. The main workflow is to acquire a token from
the SWS web interface and use the \code{\link{GetTestEnvironment}} function 
to open a session. After this, use the \code{\link{GetData}} function. To
read the vignette: \code{vignette("R_API_2014")}
}
\author{
Engineering Ingegneria Informatica
}

