% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchSWSVariable.R
\name{FetchDatatable}
\alias{FetchDatatable}
\title{Function to fetch name of current Datatable}
\usage{
FetchDatatable()
}
\description{
This function gets the name of the current Datatable that the token is
pointed toward. The internals may change at any time and it is not
recommended to access the object directly.
}

