% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchDatatableConfig.R
\name{FetchDatatableConfig}
\alias{FetchDatatableConfig}
\alias{FetchDatatableNames}
\alias{FetchDatatableNames}
\title{Read Datatable metadata list}
\usage{
FetchDatatableConfig(tables, baseurl = swsContext.baseRestUrl)

FetchDatatableNames(baseurl = swsContext.baseRestUrl)
}
\arguments{
\item{tables}{character. Optional specification of tables of interest.}

\item{baseurl}{character. If specified, uses this URL instead of looking for 
one obtained from a token. Must be of the form 
\code{http(s)://hostname:port/sws/rest}.}
}
\value{
For \code{FetchDatatableConfig}, a list containing Datatable metadata.
  For \code{FetchDatatableNames}, a character vector of available Datatables.
}
\description{
These functions pull metadata regarding Datatables. If 
  \code{\link{GetTestEnvironment}} is not called beforehand, a URL may be 
  supplied, allowing access to this metadata without a token. Of particular 
  note in each field is the 'label' element which gives the proper name of
  the table as is displayed on the web interface of the Statistical Working
  System
  
  \code{FetchDatatableNames} is a simpler version of 
  \code{FetchDatatableConfig} which gives the names of all available tables
}
