% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetBlockMetadata.R
\name{GetBlockMetadata}
\alias{GetBlockMetadata}
\title{Get Block Metadata}
\usage{
GetBlockMetadata(key)
}
\arguments{
\item{key}{A key object, as contained in a BlockMetadata object.}
}
\description{
Get Block Metadata
}

