% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCodeList.R
\name{GetCodeList}
\alias{GetCodeList}
\title{Get Code List}
\usage{
GetCodeList(domain, dataset, dimension, codes)
}
\arguments{
\item{domain}{A character value specifying the domain for which the code
list is required.}

\item{dataset}{A character value specifying the dataset for which the code
list is required.}

\item{dimension}{A character value specifying the name of the key for which
the code list is required.}

\item{codes}{(optional) A list of codes for which the key data is required.}
}
\value{
A data table containing the key codes and matching labels. Will have
  the following columns:
\itemize{
\item code - Alphanumeric code in SWS
\item description - Human-readable description
\item selectionOnly - Boolean stating whether this item is a hierarchical
category or a bottom level one containing data
\item type - Depends on dimension
\item startDate - Mainly for countries, start date for the country's existence
\item endDate - As for startData, end date of existence
}
}
\description{
This function returns information about certain codes for a dimension.  For
example, one may wish to check if a particular country has a startDate or
endDate in the system before attempting to write a particular value.  See
example below for a query that pulls the startDates for three countries.
Note that Belarus(1992-) is not defined prior to 1992.
}
\examples{
\dontrun{
GetCodeList(domain = "agriculture", dataset = "agriculture",
            dimension = "geographicAreaM49", codes = c("4", "8", "112"))
}

}

