% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCodeTree.R
\name{GetCodeTree}
\alias{GetCodeTree}
\title{Get Code Tree}
\usage{
GetCodeTree(domain, dataset, dimension, roots)
}
\arguments{
\item{domain}{A character value specifying the domain for which the code list is
required.}

\item{dataset}{A character value specifying the dataset for which the code
list is required.}

\item{dimension}{A character value specifying the name of the key for which
the code list is required.}

\item{roots}{[optional] A list of root codes for the tree. Default is all
those with no parent node.}
}
\value{
A data table containing the parent codes and lists of child codes
in the tree (see the Code Tree structure above).
}
\description{
Get Code Tree
}
\examples{
\dontrun{
GetCodeTree(domain = "agriculture", dataset = "aproduction",
                     dimension = "geographicAreaM49", roots = "953")
}

}
