% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDatasetConfig.R
\name{GetDatasetConfig}
\alias{GetDatasetConfig}
\title{Get Dataset Config}
\usage{
GetDatasetConfig(domainCode, datasetCode)
}
\arguments{
\item{domainCode}{A string containing the domain code}

\item{datasetCode}{A string containing the dataset code}
}
\value{
A object containing the full information
}
\description{
Invokes a SWS RESTful GET to query for the details of a given dataset configuration
Note: if the user of the request is not allowed for the dataset, this will not be
included in the resulting list
}

