% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDatasetNames.R
\name{GetDatasetNames}
\alias{GetDatasetNames}
\title{Get Dataset Names}
\usage{
GetDatasetNames(domainCode)
}
\arguments{
\item{domainCode}{A string containing the domain code}
}
\value{
A list of dataset names
}
\description{
Invokes a SWS RESTful GET to query for the list of all datasets in a domain
Note: if the user of the request is not allowed for a dataset, this will not be
included in the resulting list
}
\seealso{
\code{\link{GetDatasetConfig}} \code{\link{GetDomainNames}}
}
