% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDomainNames.R
\name{GetDomainNames}
\alias{GetDomainNames}
\title{Get Domain Names}
\usage{
GetDomainNames()
}
\value{
A list of domain names
}
\description{
Invokes a SWS RESTful GET to query for the list of all datasets in a domain
Note: if the user of the request is not allowed for any of the datasets in a domain, 
this domain will not be included in the resulting list
}
\seealso{
\code{\link{GetDatasetConfig}} \code{\link{GetDatasetNames}}
}

