% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlagMethod.R
\name{getFlagMethod}
\alias{getFlagMethod}
\title{Get Flag Method}
\usage{
getFlagMethod(flags)
}
\arguments{
\item{flags}{Character vector with the Faostat flag.}
}
\value{
Vector with the method flag.
}
\description{
This function converts the Faostat flag to the method flag
}
