% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlagObservationStatus.R
\name{getFlagObservationStatus}
\alias{getFlagObservationStatus}
\title{Get Flag Observation Status}
\usage{
getFlagObservationStatus(flags)
}
\arguments{
\item{flags}{Character vector with the Faostat flag.}
}
\value{
Vector with the method flag.
}
\description{
This function converts the Faostat flag to the flag osbervation status
}
