% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHistory.R
\name{GetHistory}
\alias{GetHistory}
\title{Get History}
\usage{
GetHistory(key, pivoting)
}
\arguments{
\item{key}{An object of class DatasetKey.  Often, this will be one of the
list elements of swsContext.datasets (if running in a debug/local session,
create this object with GetTestEnvironment).}

\item{pivoting}{A vector, each of whose elements must be an object of type
Pivoting.  If omitted, no pivoting is performed on the dataset.  Using this
argument can allow for convenient reshaping of the data prior to pulling it
into R.}
}
\value{
A data table of observation objects containing the values and flags
through history with the associated history metadata.
}
\description{
Get History
}

