% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLossData.R
\name{getLossData}
\alias{getLossData}
\title{Get Official Food Loss Data}
\usage{
getLossData(protected = FALSE)
}
\arguments{
\item{protected}{Logical only return observations with protected flag combination}
}
\description{
Function to obtain official food loss data at primary level
}

