% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLossData_LossDomain.R
\name{getLossData_LossDomain}
\alias{getLossData_LossDomain}
\title{Get Official Food Loss Data}
\usage{
getLossData_LossDomain(areaVar, itemVar, yearVar, elementVar, selectedYear,
  KeyEle, protected = FALSE)
}
\arguments{
\item{protected}{Logical only return observations with protected flag combination}
}
\description{
Function to obtain official food loss data at primary level
}
