% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMapping.R
\name{GetMapping}
\alias{GetMapping}
\title{Get Mapping}
\usage{
GetMapping(key)
}
\arguments{
\item{key}{A MappingTableKey class object.}
}
\value{
A data table containing the map entries matching the key (may be
empty).
}
\description{
This is expected to be used only by advanced script writers. The key will
be a custom key layout matching the axes of the mapping table.
}
\details{
The key may make use of wildcards to widen the number of results matched.

Map tables are always read-only
}

