% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNutrientData.R
\name{getNutritionData}
\alias{getNutritionData}
\title{Get Official Food Loss Data}
\usage{
getNutritionData(areaVar, itemVar, yearVar, elementVar, selectedYear,
  protected = FALSE)
}
\arguments{
\item{protected}{Logical only return observations with protected flag combination}
}
\description{
Function to obtain official food loss data at primary level
}
