% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOfficialLossData.R
\name{getOfficialLossData}
\alias{getOfficialLossData}
\title{Get Official Food Loss Data}
\usage{
getOfficialLossData()
}
\description{
Function to obtain official food loss data at primary level
}

