% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOpeningStockData.R
\name{getOpeningStockData}
\alias{getOpeningStockData}
\title{Get Opening Stock Data}
\usage{
getOpeningStockData()
}
\description{
Function to obtain opening stocks data calculated by the faoswsStock module.
}

