% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProductionData.R
\name{getProductionData}
\alias{getProductionData}
\title{Get Production Data}
\usage{
getProductionData(areaVar, itemVar, yearVar, elementVar)
}
\description{
Function to obtain production data at primary level
}
