% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/requiredItems.R
\name{getRequiredItems}
\alias{getRequiredItems}
\title{Get Required Items}
\usage{
getRequiredItems()
}
\description{
Table to obtain all primary item with CPC and FCL codes
}
\details{
Table to check
}

