% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSelectedLossData.R
\name{getSelectedLossData}
\alias{getSelectedLossData}
\title{Get Selected Food Loss Data}
\usage{
getSelectedLossData()
}
\description{
Function to obtain official food loss data at primary level
}

