% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetTestEnvironment.R
\name{GetTestEnvironment}
\alias{GetTestEnvironment}
\title{Get Test Environment}
\usage{
GetTestEnvironment(baseUrl, token)
}
\arguments{
\item{baseUrl}{The url for the SWS server, typically
"https://hqlqasws1:8181/sws" or something similar.}

\item{token}{A token which tells the SWS system what dataset to access.
This token can be obtained from the system by

\itemize{
    \item Opening a session with the relevant data.
    \item Clicking "R Plugins"
    \item Selecting the relevant script to analyze.
    \item Clicking "New debug session".
}
If the module does not currently exist on the system, you will need to
upload a zipped file with an xml file specifying the dataset
configurations.}
}
\value{
This function doesn't return any objects but creates "swsContext"
objects, such as swsContext.datasets, swsContexts.token, etc.
}
\description{
Get Test Environment
}

