% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFLI_SDG_fun.R
\name{GFLI_SDG_fun}
\alias{GFLI_SDG_fun}
\title{Part of the FAO Loss Module}
\usage{
GFLI_SDG_fun(BaseYear, keys_lower, aggregation, basket, basketKeys,
  DataForIndex)
}
\description{
Part of the FAO Loss Module
}
\author{
Alicia English Marco Migone
}
