% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleHTTPError.R
\name{HandleHTTPError}
\alias{HandleHTTPError}
\title{HTTP Error handling}
\usage{
HandleHTTPError(status, response)
}
\arguments{
\item{status}{HTTP status (numeric)}

\item{response}{Server response}
}
\description{
Gives a more verbose error message when an HTTP Error is received.
}
\author{
Sebastian Campbell <sebastian.campbell@fao.org>
}
\keyword{internal}

