% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/KeyDefinition.R
\docType{class}
\name{KeyDefinition-class}
\alias{KeyDefinition}
\alias{KeyDefinition-class}
\title{Definition for KeyDefinition Class}
\arguments{
\item{code}{A character value holding the name of the dimension in the
system (for example, "geographicAreaM49").}

\item{description}{A character value holding the description of the
dimension.}

\item{type}{A character value.  Valid values are "normal",
"measurementUnit", and "time".}
}
\description{
Definition for KeyDefinition Class
}

