% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossModel.R
\name{LossModel}
\alias{LossModel}
\title{Part of the FAO Loss Module}
\usage{
LossModel(Data, timeSeriesDataToBeImputed, production, HierarchicalCluster,
  keys_lower)
}
\description{
Part of the FAO Loss Module
}
\author{
Alicia English
}
