% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossModel_ctry.R
\name{LossModel_ctry}
\alias{LossModel_ctry}
\title{Part of the FAO Loss Module}
\usage{
LossModel_ctry(Data, timeSeriesDataToBeImputed, ctry_modelvar,
  HierarchicalCluster, keys_lower)
}
\description{
Part of the FAO Loss Module
}
\author{
Alicia English
}
