% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Metadata.R
\docType{class}
\name{Metadata-class}
\alias{Metadata}
\alias{Metadata-class}
\title{Metadata Class Definition}
\arguments{
\item{code}{A character value identifying the metadata type represented
(for instance ACCURACY or GENERAL).}

\item{description}{A character value of the description of the metadata
type.}

\item{language}{Should be one of the following character values:
\itemize{
    \item "ar": Arabic
    \item "en": English
    \item "es": Spanish
    \item "fr": French
    \item "ru": Russian
    \item "zh": Chinese
}}

\item{elements}{A list of objects of class MetadataElement.}
}
\description{
This object represents the metadata associated to the block of data
specified by the keys held by the "selection" slot for BlockMetaData.
}

