% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MetadataElement.R
\docType{class}
\name{MetadataElement-class}
\alias{MetadataElement}
\alias{MetadataElement-class}
\title{MetadataElement Class Definition}
\arguments{
\item{code}{A character value specifying the metadata element type (such as
"SAMPLING_ERROR" or "COMMENT").}

\item{description}{A character value holding the extended description of
the metadata element.}

\item{value}{The character value associated to the metadata element.}
}
\description{
MetadataElement Class Definition
}

