% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveData.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalize Data}
\usage{
normalizeData(data, keys, denormalizedKey)
}
\arguments{
\item{data}{The denormalized dataset, provided as a data.table.}

\item{keys}{A character vector of column names of data which correspond
to all the dimensions of the data except for the denormalized dimension.}

\item{denormalizedKey}{A character value containing the name of the
denormalized key.  This string is then used to determine the columns of
data which correspond to values and flags and which correspond to
dimensions.}
}
\value{
A data.table object in normalized format.
}
\description{
This function takes denormalized data and casts it as normalized data,
mostly with the help of the tidyr package.
}

