% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Pivoting.R
\docType{class}
\name{Pivoting-class}
\alias{Pivoting}
\alias{Pivoting-class}
\title{Pivoting Class Definition}
\arguments{
\item{code}{A character value containing the descriptive string of the
referred dimension (i.e. "geographicAreaM49").}

\item{ascending}{Logical, indicates the sort direction to be applied to the
specified dimension.}
}
\description{
This class represents a single dimension used for pivoting specification.
It is used in data retrieval methods to indicate the requested data
pivoting.
}
\details{
In order to properly specify a valid pivoting, a vector of Pivoting objects
has to be passed to the data retrieval code. The vector must include all
the dimensions that are defined for the target dataset. The sequence of the
Pivoting objects describe the requested order of extraction of the
corresponding dimensions and, in case of denormalized extractions, it puts
on the columns of the returned data.table the values corresponding to the
last Pivoting dimension.
}

