% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PutRestCall.R
\name{PutRestCall}
\alias{PutRestCall}
\title{Put Rest Call}
\usage{
PutRestCall(url, data)
}
\arguments{
\item{url}{The url specifying where the data should be uploaded.  Usually
this is generated by some other function, as it tends to contain
information such as the token.}

\item{data}{An object containing the data in json format.  Again, this is
usually produced by a function rather than directly supplied by the user.}
}
\value{
The response from the server.
}
\description{
Put Rest Call
}

