% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadQualityIndicators.R
\name{ReadQualityIndicators}
\alias{ReadQualityIndicators}
\title{Read Quality Indicators from dataset}
\usage{
ReadQualityIndicators(dataset)
}
\arguments{
\item{dataset}{}
}
\description{
These functions allow reading and writing of quality indicators, Quality indicators are supplied as lists.
}

