% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveBlockMetadata.R
\name{SaveBlockMetadata}
\alias{SaveBlockMetadata}
\title{Save Block Metadata}
\usage{
SaveBlockMetadata(domain, dataset, blockMetadata)
}
\arguments{
\item{domain}{A character value specifying the domain for which the code list is
required.}

\item{dataset}{A character value specifying the dataset for which the code
list is required.}

\item{blockMetadata}{An object of type BlockMetadata which should be saved
back to the user's session.}
}
\description{
Save Block Metadata
}

