% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveDatatable.R
\name{AddInsertions}
\alias{AddDeletions}
\alias{AddInsertions}
\alias{AddModifications}
\title{Save data to Datatable}
\usage{
AddInsertions(changeset, data)

AddModifications(changeset, data)

AddDeletions(changeset, data)
}
\arguments{
\item{changeset}{\code{\link{Changeset}} object}

\item{data}{data.table. 
\itemize{ 
\item AddInsertions Any __id or __ts column 
will be stripped and then sent to the server as additional lines 
\item
AddModifications __id and __ts columns must be present and all data will be
sent to the server as modifications 
\item AddDeletions Only __id and __ts
columns are requires, all others will be stripped. Removes those entries
from the table on the server
}}
}
\description{
validation hasn't been implemented yet
}
\examples{
\dontrun{
}

}

