% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveValidation.R
\name{SaveValidation}
\alias{SaveValidation}
\title{Save Validation}
\usage{
SaveValidation(domain, dataset, validation)
}
\arguments{
\item{domain}{A character value specifying the domain for which the code
list is required.}

\item{dataset}{A character value specifying the dataset for which the code
list is required.}

\item{validation}{A data.table object containing keys and validation data.}
}
\value{
The data passed to this function is saved to the database, and
nothing is explicitly returned by this function.
}
\description{
Save Validation
}

