% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SetClientFiles.R
\name{SetClientFiles}
\alias{SetClientFiles}
\title{Set Client File locations}
\usage{
SetClientFiles(certificate = "~/.R/client.crt", key = "~/.R/client.key",
  p12 = "~/.R/client.p12", p12password = "changeme",
  servercertificate = "~/.R/s1as.pem", noproxy = "*")
}
\arguments{
\item{certificate}{}
}
\description{
This function allows changing of the the locations of the client files from
their default location in '~/.R'. The client files consist of at minimum, a
client key and a client certificate. Run this command before running
\code{\link{GetTestEnvironment}} in order to set or change client file
locations. Any parameter that is not explicitly set will be reset to
defaults. As a result running \code{SetClientFiles()} will reset all
parameters to their default values.
}

