% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetTableData.R
\name{SetTableData}
\alias{SetTableData}
\title{SetTableData}
\usage{
SetTableData(schemaName, tableName, data, replace = FALSE, purge = FALSE,
  purgeFilter = NULL)
}
\arguments{
\item{schemaName}{A character value containing the name of the schema to be
accessed for the query.}

\item{tableName}{A character value containing the name of the table to be
read by the query.}

\item{data}{A non empty data.table containing records with values to be
appended/edited on the table. No need to put all the columns that are in
the DB, BUT at least the column/s that compose the primary key (if any). If
no primary key is set for the table, all the records will be appended as
they are}

\item{replace}{Logical, if TRUE, then for a row that exists in the
data.table and a corresponding record is fuond in the DB, with same key,
the record is UPDATED.  If FALSE, it is skipped.}

\item{purge}{Logical.  If true, before any inserts/updates, rows are
deleted from the database table; which rows depends on the purgeFilter
parameter (see below).}

\item{purgeFilter}{Can be specified only for purge=TRUE.  This allows the
user to specify a filter for the deletion in the form of an SQL WHERE
clause.  Specifying purge=TRUE and purgeFilter=NULL will remove all rows
from the table before the update.}
}
\description{
This function allows altering the content of a table for a specified
schema.
}
\details{
The system checks if the data passed as data.table are either
conforming to the expected column type to which any cell refers, or can be
parsed (e.g. a string containing "1" to the number 1, or a string
containing "01-01-2001" to the corresponding date by assuming the format
"dd-MM-yyyy"). If the conversion fails, or any other SQL error is arisen by
the underlying DB (e.g. a column constraint for NOT NULL), the function
aborts (using stop()) with the error message being placed on the R error
message queue.  If otherwise the function succeed, statistics on the
updates are returned, indicating how many records have been:
\itemize{
    \item deleted (if purge was TRUE)
    \item updated (if replace was TRUE and rec found by key)
    \item inserted (if rec not found by key)
    \item skipped (if replace was FALSE and rec found by key)
}
}

