% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{swsContext}
\alias{.swsenv}
\alias{swsContext}
\alias{swsContext.clientCertificate}
\title{swsContext}
\format{\preformatted{<environment: 0x00000000084a2cf8> 
}}
\usage{
.swsenv
}
\description{
These objects specify the location of important files for data access.
}
\details{
Most of these objects are created by a call to GetTestEnvironment.
}
\keyword{datasets}

