% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariablesAdd1.R
\name{VariablesAdd1}
\alias{VariablesAdd1}
\title{Part of the FAO Loss Module}
\usage{
VariablesAdd1(DataUseInt, keys_lower, Predvar2, Impute, fgroup)
}
\description{
Part of the FAO Loss Module
}
\author{
Alicia English
}
