% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteQualityIndicators.R
\name{WriteQualityIndicators}
\alias{WriteQualityIndicators}
\title{Write Quality Indicators to dataset}
\usage{
WriteQualityIndicators(indicators, dataset)
}
\arguments{
\item{indicators}{list. List of indicators to modify or add.}

\item{dataset}{character. Name of dataset to which to write Quality 
indicators}
}
\description{
Write Quality Indicators to dataset
}
\details{
The dataset may be supplied as a nested list like the \code{content}
  section received from \code{\link{ReadQualityIndicators}}. The list should 
  contain Quality Indicators that you wish to add or modify. Additions and 
  modifications occur at the top level of the list.
  
  In order to delete a value, simply have a top-level list element named as 
  NULL.
}
\examples{
\dontrun{
WriteQualityIndicators(list(a=1, b=data.frame(x=3, y="a")), "aproduction")
WriteQualityIndicators(list(a=NULL, b=NULL), "aproduction")}
  
}
\seealso{
\link{ReadQualityIndicators}
}
