% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddViolation.R
\name{AddViolation}
\alias{AddViolation}
\title{Add violation to faosws}
\usage{
AddViolation(changeset, violationtable)
}
\arguments{
\item{changeset}{\link{Changeset} object.}

\item{violationtable}{data.table with the following columns:
\itemize{
  \item __id id columns from data
  \item column character. Column name in data
  \item type character. One of 'error' or 'warning'
  \item severity integer from 1 to 10.
  \item message character. Reason for invalidating
  }}
}
\value{
Doesn't return anything, but modifies the provided changeset
}
\description{
To mark a row as valid, just return NA for everything except ID Any element
which is not otherwise marked as invalid will be considered valid. When a
changset is finalised, the whole table is considered as validated until a
cell changes value.
}
\examples{
\dontrun{
table <- "world_bank_climate_data_test"
newdat <- ReadDatatable(table, validationOptions = list(incremental = FALSE))
inval <- newdat[precipitation > temperature, .(id)]
inval[,`:=`(col = "precipitation", type="error", gravity=9, message="precipitation is greater than temperature")]
vchangeset <- Changeset(table, type = "validation")
AddViolation(vchangeset, inval)
Finalise(vchangeset)
}

}

