% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddViolation.R
\name{AddViolation}
\alias{AddViolation}
\title{Add violation to faosws}
\usage{
AddViolation(changeset, violationtable)
}
\arguments{
\item{changeset}{\link{Changeset} object.}

\item{violationtable.}{data.table with the following columns:
\itemize{
  \item __id id columns from data
  \item column character. Column name in data
  \item type character. One of 'error' or 'warning'
  \item severity integer from 1 to 10. Currently, only colours for 1 to 5 are displayed
  \item message character. Reason for invalidating}}
}
\description{
To mark a row as valid, just return NA for everything except ID
}

