% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BlockMetadataSet.R
\docType{class}
\name{BlockMetadataSet-class}
\alias{BlockMetadataSet}
\alias{BlockMetadataSet-class}
\title{The Block Metadata Set Class}
\arguments{
\item{keyDefinitions}{List of dimensions that characterize the underlying
dataset.  Each list element has to be an instance of the S4 class
KeyDefinition.}

\item{blockMetadata}{A list of BlockMetaData objects.}
}
\description{
The Block Metadata Set Class
}

