% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModifiedCells.R
\name{.buildModifiedCells}
\alias{.buildModifiedCells}
\title{Build Modified Cells}
\usage{
.buildModifiedCells(dimensions, coordinates)
}
\arguments{
\item{dimensions}{A table object of some kind}

\item{coordinates}{A vector of length equal to the number of columns of
\code{dimensions}}
}
\description{
Build Modified Cells
}

