% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveDatatable.R
\name{Changeset}
\alias{Changeset}
\alias{Finalise}
\alias{Finalize}
\title{Define changeset object}
\usage{
Changeset(table)

Finalize(changeset)
}
\arguments{
\item{table}{character. Name of Datatable.}

\item{changeset}{Changeset object.}
}
\description{
Define changeset object
}

