% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DatasetKey.R
\docType{class}
\name{DatasetKey-class}
\alias{DatasetKey}
\alias{DatasetKey-class}
\title{Dataset Key Class}
\arguments{
\item{domain}{A character value specifying the domain of interest, such as
"agriculture".}

\item{dataset}{A character value specifying the dataset within the domain,
such as "agriculture".}

\item{dimensions}{A list of objects, each of class Dimension.}

\item{sessionId}{The ID of the session from which data should be accessed.
If NULL, the database is used directly.}
}
\description{
Dataset Key Class
}
\note{
Many of these variables (domain, dataset, and dimensions) can be
found by examining some of the swsContext objects which are created in a
GetTestEnvironment call (in a debug session).  Moreover, after executing
GetTestEnvironment, the swsContext.datasets object will contain a list of
objects of type DatasetKey.
}

