% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Dimension.R
\docType{class}
\name{Dimension-class}
\alias{Dimension}
\alias{Dimension-class}
\title{Dimension Class}
\arguments{
\item{name}{The name of the dimension to access, some examples are
"timePointYears", "measuredElement", or "geographicAreaM49".  These
correspond to the dimensions on the SWS, but the names are slightly
different.  Currently, if you want to find the name of a dimension, you
need to ask CIO.}

\item{keys}{The key codes specifying what values should be pulled for the
dimension.}
}
\description{
A dimension is represented by its name and by the vector of selected
codes/keys.
}

