% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchDatatableConfig.R
\name{FetchDatatableConfig}
\alias{FetchDatatableConfig}
\title{Read Datatable metadata list}
\usage{
FetchDatatableConfig(tables, baseurl = swsContext.baseRestUrl)
}
\arguments{
\item{tables}{character. Optional specification of tables of interest.}

\item{baseurl}{character. If specified, uses this URL instead of looking for 
one obtained from a token. Must be of the form
\code{http(s)://hostname:port/sws/rest}.}
}
\value{
A list containing table metadata
}
\description{
This function pulls metadata regarding Datatables. If 
\code{\link{GetTestEnvironment}} is not called beforehand, a URL may be 
supplied, allowing access to this metadata without a token.
}

