% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetTestEnvironment.R
\name{GetTestEnvironment}
\alias{GetTestEnvironment}
\title{Get Test Environment}
\usage{
GetTestEnvironment(baseUrl, token)
}
\arguments{
\item{baseUrl}{The url for the SWS server:
\itemize{
  \item QA: \code{https://hqlqasws1.hq.un.fao.org:8181/sws}
  \item Production: \code{https://hqlprswsas1.hq.un.fao.org:8181/sws}
}}

\item{token}{A token which tells the SWS system what dataset to access. This
  token can be obtained from the system by

\itemize{
    \item Opening a session with the relevant data.
    \item Clicking "R Plugins"
    \item Selecting the relevant script to analyze.
    \item Clicking "New debug session".
}

If the module does not currently exist on the system, you will need to
upload a zipped file with an xml file specifying the dataset configurations.}
}
\value{
This function silently returns a list containing a DatasetKey object and creates
  "swsContext" objects, such as \code{swsContext.datasets}, \code{swsContexts.token},
  etc., including \code{swsContext.datasets} which is identical to the object
  returned.
}
\description{
Get Test Environment
}

