% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappingTableKey.R
\docType{class}
\name{MappingTableKey-class}
\alias{MappingTableKey}
\alias{MappingTableKey-class}
\title{Mapping Table Key}
\description{
Mapping Table Key
}
\section{Slots}{

\describe{
\item{\code{dimensions}}{}
}}

