% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataElement.R
\docType{class}
\name{MetadataElement-class}
\alias{MetadataElement}
\alias{MetadataElement-class}
\title{MetadataElement Class Definition}
\description{
MetadataElement Class Definition
}
\section{Slots}{

\describe{
\item{\code{code}}{A character value specifying the metadata element type (such as
"SAMPLING_ERROR" or "COMMENT").}

\item{\code{description}}{A character value holding the extended description of
the metadata element.}

\item{\code{value}}{The character value associated to the metadata element.}
}}

