% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadQualityIndicators.R
\name{ReadQualityIndicators}
\alias{ReadQualityIndicators}
\title{Read Quality Indicators from dataset}
\usage{
ReadQualityIndicators(dataset)
}
\arguments{
\item{dataset}{character. Name of an SWS dataset}
}
\value{
A list with the following structure
  \itemize{
    \item \code{id} Unique id of this version of indicators
    \item \code{version} Version of the schema to which indicators adhere
    \item \code{model} Name of the schema, in this case always 'sws-indicators'
    \item \code{target} The metadata target. For our purposes, the dataset name
    \item \code{content} A nested list containing the quality indicators
    \item \code{createdOn} Unix timestamp of point where indicators were first 
      created
    \item \code{lastModifiedOn} Unix timestamp of point when indicators were 
      last modified
    \item \code{createdBy} User who first created indicators
    \item \code{lastModifiedBy} User who last modified indicators
  }
}
\description{
These functions allow reading and writing of quality indicators, Quality
indicators are supplied as lists.
}

