% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveDatatable.R
\name{AddInsertions}
\alias{AddDeletions}
\alias{AddInsertions}
\alias{AddModifications}
\alias{SaveDatatable}
\title{Save data to Datatable}
\usage{
AddInsertions(changeset, data)

AddModifications(changeset, data)

AddDeletions(changeset, data)
}
\arguments{
\item{changeset}{\code{\link{Changeset}} object}

\item{data}{data.table. 
\itemize{ 
\item AddInsertions Any __id or __ts column 
will be stripped and then sent to the server as additional lines 
\item
AddModifications __id and __ts columns must be present and all data will be
sent to the server as modifications 
\item AddDeletions Only __id and __ts
columns are required, all others will be stripped. Removes those entries
from the table on the server
}}
}
\description{
Functions to save to Datatables on SWS.
}
\examples{
\dontrun{
library(data.table)
# define original Datatable
origtable <- "world_bank_climate_data"
# define test table to write to
table <- "world_bank_climate_data_test"

# Read new table
newdat <- ReadDatatable(table, readOnly = FALSE)
# Read original table
dat <- ReadDatatable(origtable, limit=10)

## Erase all data in test table
# Define changeset object
changeset <- Changeset(table)
# Add rows to delete to changeset
AddDeletions(changeset, newdat)
# Send modifications to the server
Finalise(changeset)

# Add new data (twice) to test from original
AddInsertions(changeset, dat)
AddInsertions(changeset, dat)
Finalise(changeset)

# Reread test table
newdat <- ReadDatatable(table, readOnly = FALSE)

# Double precipitation in first half of table
newdat[1:ncol(dat),precipitation := precipitation * 2]
AddModifications(changeset, newdat[1:nrow(dat),])

# Delete the second half of the table
AddDeletions(changeset, newdat[nrow(dat):(nrow(dat)*2),])
# Send it back to the server
Finalise(changeset)
}

 
}

