% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetClientFiles.R
\name{SetClientFiles}
\alias{SetClientFiles}
\title{Set Client File locations}
\usage{
SetClientFiles(dir = "~/.R", certificate = file.path(dir, "client.crt"),
  key = file.path(dir, "client.key"), p12 = file.path(dir, "client.p12"),
  p12password = "changeme", servercertificate = file.path(dir, "s1as.pem"),
  noproxy = "*")
}
\arguments{
\item{dir}{Directory where the client certificates are installed}

\item{certificate}{Path to "client.crt" file}

\item{key}{Path to "client.key" file}

\item{p12}{path to "client.p12" file}

\item{p12password}{Password relating to logging in using Mac?}

\item{servercertificate}{See above.}

\item{noproxy}{?}
}
\description{
This function allows changing of the the locations of the client files from 
their default location in '~/.R'. The client files consist of at minimum, a 
client key and a client certificate. Run this command before running
\code{\link{GetTestEnvironment}} in order to set or change client file
locations. Any parameter that is not explicitly set will be reset to
defaults. As a result running \code{SetClientFiles()} will reset all
parameters to their default values.
}
\author{
Sebastian Campbell
}

