% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{swsContext}
\alias{swsContext}
\alias{swsContext.clientCertificate}
\title{swsContext}
\format{\preformatted{ chr "~/.R/client.crt"
}}
\usage{
swsContext.clientCertificate
}
\description{
These objects specify the location of important files for data access.
}
\details{
Most of these objects are created by a call to GetTestEnvironment.
}
\keyword{datasets}

